Add/Update/Delete Departments working through web page

+++ src/main/java/com/systemsinmotion/orgchart/service/DepartmentService.java
@@ -1,6 +1,7 @@
 package com.systemsinmotion.orgchart.service;
 
 import java.util.List;
+import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -19,11 +20,28 @@ public class DepartmentService {
 	}
 
 	public Department findDepartmentByID(Integer departmentId) {
-		return this.repository.findOne(departmentId);
+		return this.repository.findById(departmentId);
+	}
+	
+	public Department findDepartmentByName(String departmentName){
+		return this.repository.findByName(departmentName);
 	}
 
 	public void removeDepartment(Department department) {
-		this.repository.delete(department);
+		if(department != null){
+			List<Department> depts = findAllDepartments();
+			for(Department child : depts){
+				Department parent = child.getParentDepartment();
+				if(parent != null)
+					if(parent.getId().equals(department.getId())){
+						child.setParentDepartment(department.getParentDepartment());
+						storeDepartment(child);
+					}
+			}
+			department.setParentDepartment(null);
+			storeDepartment(department);
+			this.repository.delete(department);
+		}
 	}
 
 	public void setRepository(DepartmentRepository repository) {
@@ -31,7 +49,11 @@ public class DepartmentService {
 	}
 
 	public Department storeDepartment(Department department) {
-		return this.repository.save(department);
+		if(findDepartmentByName(department.getName()) != null && department.getId() == null)
+			return null;
+		Department parent = department.getParentDepartment();
+		if(parent != null)
+			department.setParentDepartment(parent.getId() == null ? null : parent);
+		return this.repository.saveAndFlush(department);
 	}
-
 }
